downloads: make ask_for_filename accept a callback
Rather than trying to mix two styles here (blocking-and-return and async
callbacks), it's probably easier to settle for one. This eliminates the
need to return the (filename, question)-tuple.

Because it's easier to convert to the callback style, ask_for_filename
now always uses a callback instead of returning the filename. The
question is still returned because the download-manager might need it.