qutescheme: add handler for qute://
A test uses qute:// and this test fails if the download produces an
error ("No handler found for qute://").

The test worked earlier because the file was not actually downloaded,
only the filename-ask-function was called. Since this was changed, we
now need some "content" at qute:// to make the download succeed.